ml/requirements.txt
unsloth>=2025.1.0
torch>=2.4
transformers>=4.46
trl>=0.12
peft>=0.13
accelerate>=1.0
bitsandbytes>=0.44
datasets>=3.0
Pillow>=10.0
opencv-python>=4.9
shapely>=2.0
tqdm>=4.66
numpy>=1.26
httpx>=0.27
